Raise an exception on empty products in Python.
The HARP command line tools treat an empty product as an error, whereas the
Python interface just returned an empty instance of harp.Product. To make the
behaviour more consistent, the Python interface was changed to raise a
NoDataError exception on empty products. This exception is raised if the HARP
C library returns an empty product, either during ingestion or import, after
all actions have been executed on the product.