Use public API for executing actions in harpfilter.
The harpfilter command-line tool used several functions declared in
harp-internal.h to execute actions on a product. This does not play nice with
Windows builds.

Recently, a function to execute actions on a product was added to the public
API. This commit replaces several calls to functions declared in
harp-internal.h by a single call to the newly added function,
harp_product_execute_actions().