Add bit mask test operators
Added operators =& and !& to test integer variables against a bit mask, where
a =& mask is equivalent to (a & mask) != 0, i.e. at least one of the bits set
in mask is set in a, and a !& mask is equivalent to (a & mask) == 0, i.e. none
of the bits set in mask are set in a.