Make harp-ingestion action execution must stricter
The action evaluators used to be very optimistic; assuming that if they
failed, some other evaluator might still succeed.
This made error reporting unprecise, as we can only detect errors after
all evaluation functions ran and actions remained in the action list.

In the new situation, action evaluators know that if they encounter an
action of a certain type, they *have* to evaluate it.
If any error condition occurs, this can be propagated, leading to
precise error messages.