Add support for updating 2-D dimension masks from 1-D variable filters
When evaluating variable value filters during ingestion it is assumed that the
dimensionality of the dimension mask matches the dimensionality of the
variable. This is ensured by first filtering all 0-D variables, then all 1-D
variables, and finally all 2-D variables.

With user defined action ordering, several blocks of variable filters may have
to be executed. Thus, it may happen that the current mask for a secondary
dimension is 2-D when filtering a 1-D variable. (This is currently the only
case that is problematic).

This commit provides support for this problematic case by creating a temporary
1-D mask from the 2-D mask, execute the value filters on the 1-D variable, and
then propagate the changes from the 1-D mask to the 2-D mask.