Change valid range filter into a value filter
The valid range filter used to be a special type of value filter because it
could be applied to any n-d variable as long as its outer dimension was the
time dimension. If all values in the sub-array corresponding to a certain index
along the time dimension would be invalid, this index would be masked.

This functionality was a remnant of GECA which is no longer needed for HARP.
Therefore, the implementation was changed such that the valid range filter is
evaluated like any other value filter. This means it can only be applied to
variables with two dimensions or less, none of which are independent
dimensions, and if the variable is two-dimensional the outer dimension should
be the time dimension.