Allow 'molar mass' as a variable
Even though this quantity will likely never be stored explicitly in products,
it is the main quantity that is needed to perform 'wet air' calculations.
Introducing it as a variable will allow variable conversions to retrieve the molar mass
of total air from any of the available methods (pressure+nd, H2O mmr, H2O vmr, ...).