Add variable conversion for datetime_length from array to scalar
If the last dimension of datetime_length is the spectral or vertical dimension
then allow derivation of a variant without that dimension by taking the minimum value of datetime_length (i.e. the minimum time range that is applicable to
all measurments).
This assumes that datetime_start and datetime_start cannot be dependent on the spectral or vertical dimension.