Fix problem in raising error
arg[0] is an integer for CLibraryError exceptions, so we cannot add a string value to it.
It is generally incorrect to assume that the first argument to a raised exception is the error message.
We therefore raise a new exception and embed the raised exception string in the new exception.