Add support for 'missing' lowest layer for vertical data in S5P L2 O3 column product
For non-cloudy pixels the lowest layer is zero.
For L2 processor <= 01.01.02 this was not handled correctly, and for later versions the lowest layer is set to an empty length (i.e. first two pressure level values are the same).

In HARP we now (both for old and new processor versions) import the profile with one less layer in those cases (i.e. skip zero-length lower layer and use NaN fill value at the end).