Allow lat/lon bounds {..,2} and add missing geometry libharp docs
We now support operations to work on spatial boundaries defined by bounding rects.
If the independent dimension of latitude_bounds/longitude_bounds is 2 then it is a bounding rect.
If it is higher, then the variables define the vertices of a bounding polygon.

Also add the missing libharp documentation on the public geometry functions.