Fix upper bound extrapolation issue in deriving axis bounds when using padding
In case an axis variable is time dependent then the extrapolation rule should be performed on the last _valid_ axis element, which does not have to be the last element of the array.
We therefore now ignore all trailing NaN values in the axis array (and set corresponding bound values to NaN as well) when deriving an <axis>_bounds from a <axis> variable.