Speed up variable derivation
The previous introduction of a penalty based search for variable derivations introduced a large perfomance penalty when searching for derivations. This is most notable when performing a 'harpdump --list-derivations <product>'
Performance has been improved somewhat (but not to orginal levels) by comparing the current penalty during a search against the best solution thusfar.