Change HCHO column AVK to tropospheric column AVK
The HCHO column AVK was already scaled to the tropospheric AMF by the processor, so it has alsways been a tropospheric AVK.
The variable should thus have been called tropospheric_HCHO_column_number_density_avk from the start as well.
As of processor version 2.0.0 we now also cut off the AVK at the tropopause based on the new tropopause index (by setting higher values to 0 as we do for the tropospheric NO2 AVK).