Change bin() and bin_spatial() operations to use count/weight variables
The count and weight variables are used to allow the following commands to now be equivalent:

    harpmerge -a 'bin()' -ap 'bin()' a.nc b.nc merged.nc
    harpmerge -ap 'bin()' a.nc b.nc merged.nc

As well as the following:

    harpmerge -a 'bin_spatial(...)' -ap 'squash(time, (latitude_bounds, longitude_bounds));bin()' a.nc b.nc merged.nc
    harpmerge -ap 'bin_spatial(...)' a.nc b.nc merged.nc

In other words, if you bin the concatenation of products that were individually binned, this should have the same result as if you would bin the concatenation of the original products.

For averaging of angles, the weight variable is now used to store the norm of the complex vector that is used to calculate the angle average.

This commit also removes a few private functions from the API documentation.