Fix issue in udunits2 initialisation if the wrong locale was set.
The problem is that udunits2 uses strtod() to interprets values from its xml configuration files.
Since this is locale aware it will fail if the LC_NUMERIC is set to a locale that does not use '.' as a decimal separator.
This is fixed by explicitly setting the locale to a compatible one before udunits2 initialization (and resetting it afterwards).

We also set the locale explicitly for a use of strtod() in 'harpcollocate matchup' that deals with parameter parsing.