Fix issues in molar mass calculation for species
- the molar mass for variables with a 'surface', 'tropospheric', or 'stratospheric' prefix returned 0.
This was because the detection of the species name only happened at the start of a variable name.
We now first skip these known prefixes before determining the species name from a variable name.

- the molar mass was incorrectly provided for the species with the shortest matching name.
e.g. the molar mass of CO was returned for CO2 and of NO for NO2.
We now compare against the whole species name (by checking for the trailing '_') to determine a match.