Fix issue in weightfactor calculation of bin_spatial()
The issue happened when the grid resolution was about the same, or higher, than the resolution of the input data.
For grid cells that lie mostly within the polygon of the input product but that still interesected with the polygon edges the weight factor could end up higher than 1.0 (but always less than 2.0).