Fix issue in creation of missing count/weight variables when merging products
For count/weight variables with sub-dimensions, the missing variable was created using the sub-dimension length of the other product.
This would go wrong if both products didn't have the same length for that dimension (e.g. not the same number of vertical levels).
The missing variable is now created using the length of the that dimension of the current product.