Add support for revised NRTI indication
UPAS products used to use NRTI as value for /METADATA/GRANULE_DESCRIPTION@ProcessingMode, which was not in line with the specification. With the new processor version this attribute will contain the correct value 'Near-realtime' in case of NRTI.
HARP now interprets products with either NRTI or Near-realtime for the this attribute as NRTI products.
Note that this only impacts the ingestions of HCHO, O3, and SO2 (where the format between NRTI and OFFL differs).