Fixed issue where 'index(dimension)' operation would not get applied correctly.
This issue would only occur when using the filter as part of ingesting a product via an ingestion module (i.e. it was not stored in HARP format) and there was also anoter filter provided.
If the other filter introduced a 2d mask on for the given dimension (which can only happen if the filter dimension is not the time dimension), then the index filter would only get applied to the first time element (instead of globaly).