Add optimization for finding derivations based on chemical species
The majority of derivations are for variables linked to a specific chemical species.
And the majority of those can only be performed if there is already a variable for that species in the product.
This change introduces a mechanism where, if we are looking for source variables for a derivation, we directly mark the source variable as impossible if there was no variable in the product for the same species and the variable could only have been derived that way.

This especially solves the problem with the recently added derivation for (dry) air column number density from species column number density and species (dry air) volume mixing ratio, and where trying to look for a derivation for the air number density had become very slow.

With this change the maximum search depth is changed back from 8 to 10 again.